pyproject.toml
[tool.poetry]
name = "arfa-med[api]"
version = "0.1.0"
description = ""
authors = ["Janik Tarverdyan <Janik.Tarverdyan@gmail.com>"]
license = "GPLv3"
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"

django-jazzmin = "^2.6.0"
django-rest = "^0.8.7"
django-ckeditor = "^6.7.0"
django-rest-framework = "^0.1.0"
django-modeltranslation = "^0.18.11"

django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.2"
pillow = "^10.2.0"



[tool.poetry.group.dev.dependencies]
ipython = "^8.13.0"
pep8 = "^1.7.1"
autopep8 = "^2.0.2"
rich = "^13.3.5"
typing-extensions = "^4.8.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
